chore: use a dedicated file for the sub command
